refactor: move WebGL logic from HomeComponent to HomeglComponent
- Created standalone HomeglComponent for WebGL background rendering
- Migrated all WebGL methods (init, shader management, rendering loop)
- Integrated <app-homegl> into HomeComponent
- Cleaned up unused methods and fixed typos in migrated code
