+++ frontend/src/app/home/components/home/home.component.css
@@ -7,18 +7,3 @@
   height: 100%;
   z-index: 10;
 }
-
-.gl-canvas {
-    opacity: 0.8;
-  position: absolute;
-  top: 20%;
-  left: 0%;
-
-  width: 100%;
-  height: 60%;
-  z-index: 0;
-}
-
-
-
-
