+++ frontend/src/app/home/components/home/home.component.html
@@ -1,5 +1,4 @@
 
-
 <main class="flex flex-col h-full" itemscope itemtype="https://schema.org/WebApplication">
   <meta itemprop="name" content="PianoML">
   <meta itemprop="applicationCategory" content="EducationalApplication">
@@ -12,7 +11,7 @@
   
   <section id="section1" class="section1" aria-labelledby="main-heading">
 
-    <canvas #glCanvas class="gl-canvas" aria-hidden="true"></canvas>
+    <app-homegl></app-homegl>
 
 
     <div class="absolute inset-0 z-1 flex flex-col items-center justify-center gap-2">
