+++ frontend/src/app/home/components/homegl/homegl.component.css
@@ -1,0 +1,9 @@
+.gl-canvas {
+  opacity: 0.8;
+  position: absolute;
+  top: 20%;
+  left: 0%;
+  width: 100%;
+  height: 60%;
+  z-index: 0;
+}
