+++ frontend/src/app/home/components/homegl/homegl.component.html
@@ -1,0 +1,1 @@
+<canvas #glCanvas class="gl-canvas" aria-hidden="true"></canvas>
