+++ frontend/src/app/home/components/homegl/homegl.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { HomeglComponent } from './homegl.component';
+
+describe('HomeglComponent', () => {
+  let component: HomeglComponent;
+  let fixture: ComponentFixture<HomeglComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [HomeglComponent, HttpClientTestingModule]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(HomeglComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
